Add files via upload
This is a solution that includes the Mixed Reality control inside a canvas app, with a Dataverse backend to store the measurements and photos taken in the control.  There is a Power Automate flow that handles the writing of the data to Dataverse when the user clicks the Save button in the canvas app.